+++ CHANGELOG.md
@@ -45,7 +45,9 @@ for soon-to-be removed features.
 
 ### Fixed
 
-- **select**: only one **select** will remain expanded at a time, even when multiple selects are present on the page. Previously, when multiple instances of the **select** were present on a page, they could both be expanded simultaneously, which was not intended behavior.
+- **select**: 
+  - only one **select** will remain expanded at a time, even when multiple selects are present on the page. Previously, when multiple instances of the **select** were present on a page, they could both be expanded simultaneously, which was not intended behavior.
+  - add keyboard navigation and aria to respect listbox pattern
 - **doc**: add the SNCF font icon in the stylesheet on the integration doc for angular.
 - **doc**: the events are shown back in the Actions tab.
 - **doc**: storybook controls improvements for stories + attributes now displayed in table